+++ project/bl/feed.py
@@ -1,0 +1,9 @@
+from project.bl.utils import BaseBL
+
+
+class CategoryBL(BaseBL):
+    pass
+
+
+class VacancyBL(BaseBL):
+    pass
