+++ project/database.py
@@ -15,5 +15,5 @@ Base.query = db_session.query_property()
 def init_db():
     Base.metadata.drop_all(bind=engine)
     import project.auth.models  # NOQA
-    import project.feed.models  # NOQA
+    import project.models  # NOQA
     Base.metadata.create_all(bind=engine)
